requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-groq>=0.0.1
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain-core>=0.1.0
